[cpu] Don't reuse shuffle dummies (#88)
This results in the following compile-time assertion error (in debug Triton builds):

    Assertion `Index < size() && "invalid index for value range"' failed.

This occurs when there is more than one tt.reduce call with a given
number of arguments in the same function, with the later call using more
arguments. Reusing the dummy values means that the subsequent call has
fewer dummy values than expected, hence the error.

This bug also resulted in type mismatches errors between the reused
dummy value and the current input value.