Implement device_assert (#126)
This implements device_assert using a similar approach to @minjang's
implementation of device_print -- we create a new
triton_cpu.assert op that takes a vector which we then lower to our own
`triton_assert` implementation.

One notable difference between our implementation vs the GPU one is that
`abort()` on the CPU aborts the entire program, not just a single
thread. In contrast, `__assert_fail` on the GPU seems to stop execution
of the current thread, but other threads still continue execution. Thus
we only get ~one assert message even if the assert condition is false
for multiple threads, while the same program on the GPU gives us assert
messages for all threads.

I suppose there are some ways around this, but for now, let's go with
the simplest implementation.