Use 1-D vector reduction op to convert reduce op (#152)
Add `useReductionOp` option for reduce op lowering. The
`mapToMultiDimReductionOp` is renamed into `mapToReductionOp`, When
the src is a 1-D vector, replace `triton::ReduceOp` with
`vector::ReductionOp`.

Currently, `triton::ReduceOp` is lowered in a generic way with shuffle
and accumulate. However, some architecture support dedicated reduction
instructions (e.g. `vredsum` in RVV), which cannot be inst-selected
from current shuffle + accumulate pattern, and sub-optimal code might
be generated. The 1-D reduction can be converted later for specific
target in the Target TTCIR stage, or just leave it to LLVM for code
generation.

An alternate approach is keeping the `mapToMultiDimReductionOp`
enabled by default, and let the multi-reduction be converted to
`llvm.vector.reduce.*` intrinsics. But according to the previous PR,
current code generation can lead to better results for 2-D reductions.

Signed-off-by: Junyi Mei <junyi.mei@terapines.com>