Lower memory ops with vector gather and scatter
This commit add lowerToGather and lowerToScatter for load and store
conversion. Memory ops with the pointer computed from splat and addptr
can be lowered with vector.gather or vector.scatter.

For architectures with scatter and gather support (like SVE and RVV),
the code generated with this approach might be more efficient.

Two options are added to scalarization and memory op conversion to
enable lowering with gather and scatter operations.

Signed-off-by: Junyi Mei <junyi.mei@terapines.com>