Pad size 2 vectors to size 4 when lowering extern_elementwise ops
libsleef does not implement 2-element functions, so for libdevice
functions implemented via extern_elementwise ops that rely wholly on
libsleef implementations (as opposed to MathOps which can be lowered to
native instructions), we need to pad those vectors to size 4.

This allows us to enable test_math.py for all the functions introduced
in https://github.com/triton-lang/triton-cpu/pull/161.