Fix math tests for armv8 (#178)
We only use Sleef on !x86 platforms. Sleef APIs are not fully agnostic
of the underlying architecture. For example, `Sleef_sinf8_u10` does not
exist on Arm.

This PR, makes the `MathToVecLibPass` aware of the
CPU SIMD architecture by accepting `cpu_features` as new optional
argument.

No change is expected on x86 side.